requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
fp-NGFW-SMC-python>=1.0.32
cryptography>=42.0
pyyaml
gunicorn>=22.0
authlib>=1.3
requests>=2.31
